+++ mathforge/output/__init__.py
@@ -1,4 +1,15 @@
 from mathforge.output.answer_validator import AnswerValidator
 from mathforge.output.deterministic_formatter import DeterministicFormatter
+from mathforge.output.gradeability import (
+    ScorerRoundTrip,
+    enforce_scorer_round_trip,
+    is_gradeable_response,
+)
 
-__all__ = ["AnswerValidator", "DeterministicFormatter"]
+__all__ = [
+    "AnswerValidator",
+    "DeterministicFormatter",
+    "ScorerRoundTrip",
+    "enforce_scorer_round_trip",
+    "is_gradeable_response",
+]